move getQueryClient to features/shared/query-client
Fixes import hierarchy violation where features imported from the app
root. All consumers now import from ~/features/shared/query-client.
