+++ app/entry.client.tsx
@@ -10,8 +10,8 @@ import { hydrateRoot } from 'react-dom/client';
 import { HydratedRouter } from 'react-router/dom';
 import { getEnvironment, isServedLocally } from './environment';
 import { featureFlagsQueryOptions } from './features/shared/feature-flags';
+import { getQueryClient } from './features/shared/query-client';
 import { Money } from './lib/money';
-import { getQueryClient } from './query-client';
 import { getTracesSampleRate, sanitizeUrl } from './tracing-utils';
 
 // Register Chrome DevTools custom formatter for Money class (dev only)
