+++ app/routes/_auth.tsx
@@ -1,7 +1,7 @@
 import { Outlet, redirect } from 'react-router';
 import { LoadingScreen } from '~/features/loading/LoadingScreen';
+import { getQueryClient } from '~/features/shared/query-client';
 import { authQueryOptions } from '~/features/user/auth';
-import { getQueryClient } from '~/query-client';
 import type { Route } from './+types/_auth';
 
 const routeGuardMiddleware: Route.ClientMiddlewareFunction = async (
