+++ app/routes/api.lnurlp.callback.$userId.ts
@@ -5,8 +5,8 @@
 
 import { agicashDbServer } from '~/features/agicash-db/database.server';
 import { LightningAddressService } from '~/features/receive/lightning-address-service';
+import { getQueryClient } from '~/features/shared/query-client';
 import { Money } from '~/lib/money';
-import { getQueryClient } from '~/query-client';
 import type { Route } from './+types/api.lnurlp.callback.$userId';
 
 export async function loader({ request, params }: Route.LoaderArgs) {
